Refactor `IdOrdEqHash` and `Filter` into separate files
