+++ data_model/src/lib.rs
@@ -165,11 +165,10 @@ pub enum Parameter {
     Debug,
     Display,
     Clone,
-    IdOrdEqHash,
-    // PartialEq,
-    // Eq,
-    // PartialOrd,
-    // Ord,
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
     Decode,
     Encode,
     Deserialize,
@@ -178,7 +177,6 @@ pub enum Parameter {
     IntoSchema,
 )]
 #[allow(clippy::enum_variant_names)]
-#[id(type = "Self")]
 pub enum IdBox {
     /// [`DomainId`](`domain::Id`) variant.
     DomainId(<domain::Domain as Identifiable>::Id),
